[Vertical Tabs] Reposition Context Menu
Shift the context menu container view to the right by the width
of Vertical Tabs when it is on.

Before: https://screenshot.googleplex.com/373q5xaRFGrN4kd
 After: https://screenshot.googleplex.com/5LGCEwcMKaVd9V9

Bug: 521986417
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/8008621
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1653496}

+++ components/embedder_support/android/junit/src/org/chromium/components/embedder_support/contextmenu/ContextMenuUtilsUnitTest.java
@@ -575,7 +575,7 @@ public class ContextMenuUtilsUnitTest {
 
         int triggeringTouchXDp = 100;
         int triggeringTouchYDp = 200;
-        float topContentOffsetPx = 50f;
+        int topContentOffsetPx = 50;
 
         ContextMenuParams params =
                 new ContextMenuParams(
@@ -604,6 +604,7 @@ public class ContextMenuUtilsUnitTest {
                         context,
                         mockWindow,
                         params,
+                        /* leftContentOffsetPx= */ 0,
                         topContentOffsetPx,
                         isPopup,
                         mockContainerView);
@@ -682,7 +683,7 @@ public class ContextMenuUtilsUnitTest {
         View containerView = new View(context);
         int triggeringTouchXDp = 100;
         int triggeringTouchYDp = 200;
-        float topContentOffsetPx = 50f;
+        int topContentOffsetPx = 50;
 
         int shadowWidth = 60;
         int shadowHeight = 50;
@@ -717,6 +718,7 @@ public class ContextMenuUtilsUnitTest {
                         context,
                         window,
                         params,
+                        /* leftContentOffsetPx= */ 0,
                         topContentOffsetPx,
                         isDragDropEnabled /*usePopupWindow but should not matter for this case*/,
                         containerView);
